Add test modules
