+++ tests/base/module_3.py
@@ -1,8 +1,4 @@
-from src.decorator import load_config
-
-
-@load_config(order=1)
-class TestA2:
+class Module3:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
