+++ tests/base/packageC/module_c1.py
@@ -1,4 +1,8 @@
-class TestA3:
+from src.decorator import load_config
+
+
+@load_config(order=3)
+class ModuleC1:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
