+++ tests/base/packageC/module_c2.py
@@ -1,4 +1,8 @@
-class TestB2:
+from src.decorator import load_config
+
+
+@load_config(order=2)
+class ModuleC2:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
