+++ tests/base/packageC/module_c3.py
@@ -1,4 +1,8 @@
-class TestA1:
+from src.decorator import load_config
+
+
+@load_config(order=1)
+class ModuleC3:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
