+++ tests/packageA/module_a1.py
@@ -1,4 +1,4 @@
-class Test3:
+class ModuleA1:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
