+++ tests/packageA/module_a2.py
@@ -1,8 +1,8 @@
 from src.decorator import load_config
 
 
-@load_config(order=2)
-class TestC2:
+@load_config(order=1)
+class ModuleA2:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
