+++ tests/packageA/module_a3.py
@@ -1,4 +1,4 @@
-class Test2:
+class ModuleA3:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
