+++ tests/packageA/packageB/module_b1.py
@@ -1,8 +1,4 @@
-from src.decorator import load_config
-
-
-@load_config(order=3)
-class TestC1:
+class ModuleB1:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
