+++ tests/packageA/packageB/module_b2.py
@@ -1,8 +1,4 @@
-from src.decorator import load_config
-
-
-@load_config(order=1)
-class TestC3:
+class ModuleB2:
     def __init__(self):
         print(f'{self.__class__.__name__} init!')
